delete trash files, add utils js"
