--- js/getRandomArrayElement.js
@@ -1,6 +1,0 @@
-import {getRange} from './getRandomNumb.js';
-
-const getRandomArrayElement = (elements) =>
-  elements[getRange(0, elements.length - 1)];
-
-export {getRandomArrayElement};
